fix potential bug with boost::thread being terminated
in recent versions of boost::thread the behavior of its destructor was
adapted to the one of std::thread. It will now call std::terminate
instead of join() or detach() as it was done before. Therefore an extra
detach() was added so it is guaranteed to actually execute the
background job.

Also moved all boost::this_thread::sleep to std::this_thread::sleep_for
in this file